Split CsonSerializer into partial classes, to colocate Enum, List, Pairs, and Properties related code.
