+++ src/Fixie.Assertions/CsonException.cs
@@ -1,0 +1,9 @@
+﻿namespace Fixie.Assertions;
+
+class CsonException : Exception
+{
+    public CsonException(string message)
+        : base(message)
+    {
+    }
+}
