+++ src/Fixie.Assertions/CsonSerializer.Pairs.cs
@@ -1,0 +1,46 @@
+﻿namespace Fixie.Assertions;
+
+partial class CsonSerializer
+{
+    static Type? GetPairType(Type typeToConvert)
+    {
+        var enumerableType = GetEnumerableType(typeToConvert);
+            
+        if (enumerableType != null)
+        {
+            var itemType = enumerableType.GetGenericArguments()[0];
+
+            if (itemType.IsGenericType &&
+                itemType.GetGenericTypeDefinition() == typeof(KeyValuePair<,>))
+                return itemType;
+        }
+
+        return null;
+    }
+
+    static void WritePairsLiteral<TKey, TValue>(CsonWriter writer, IEnumerable<KeyValuePair<TKey, TValue>> value)
+    {
+        writer.WriteStartObject();
+
+        bool any = false;
+        foreach (var item in value)
+        {
+            if (!any)
+            {
+                writer.StartItems();
+                any = true;
+            }
+            else
+                writer.WriteItemSeparator();
+
+            writer.WritePropertyName(item.Key?.ToString()!);
+
+            SerializeInternal(writer, item.Value);
+        }
+
+        if (any)
+            writer.EndItems();
+
+        writer.WriteEndObject();
+    }
+}
